feat: html and text helper
